Mattias/testing the novac ppp (#23)
* Resolved build failures on Ubuntu 22 and resolved some warnings

* Corrected failing tests

* Corrected test which now failed on windows

* Moved files from PPPExe to PPPLib, for testability. Cleanup of headers

* Resolved build issues on Linux

* resolved warnings and remvoed unused files

* Moved classes from PPPExe to PPPLib for testability

* Split out functionailty from PostEvaluationController and tested

* Minor cleanup

* const correctness

* Started adding tests to ScanEvaluation and added verification of the input

* More tests on ScanEvaluation

* Some more tests and refactorings of the scan evaluation

* Separated out some more code from the PostEvaluationController

* Testing of running evaluation on scan with calibrated references

* Added unit tests to the EvaluationLogFileHandler, and changed signatures.

* Updated tests on evaluation log file handler

* Added tests on flux calculation and reading of the evaluation logs

* Linux fixes

* Use of log and member UserSettings in ftpServer

* Improved the logging

* Updated logging and error handling

* Log improvement and correction of one crash

* Added log context and standard strings and vectors instead of odd custom containers

* Log improvements

* Corrected a bug when determining optimium shift/squeeze from fraunhofer ref

* Updated argumetn handling

* Debug output

* Updated spectral eval

* Corrected error in skipping the evaluations, no plume properties were calculated. Corrected bug in applying command line options

* Corrected an error with the New Zeeland volcanoes, they were located on the northern hemisphere. Minor refactoring of the volcano structure

* Misc utils and settings

* cleanup of geometry calculation and cleaned up code to avoid bugs

* Corrected errors in geometry calculations

* Cleanup and renaming of FluxResult and WindField

* Rename

* Added nullable and made use of this new type

* Rename of structure

* moved exception to SpectralEvaluation, cleanup of PlumeDatabase

* Cleanup of the flux calculation

* Updated test cases

* Corrected linux problems and warnings

* Separated out the chcking of the evaluation references to a separate function

* Clarified the unit when filtering refernece

* Moved enum to SpectralEvaluation and updated after changed casing

* Updated the way the plume properties are calculated

* Corrected one bug in ThreadUtils::GuardedList which caused an error in the post processing

* Moved ThreadUtils to SpectralEvaluation

* Removed gps functions which already exist in SpectralEvaluation

* Updated spectralEvaluation

* Corrected log output

* Added signed/unsigned type check

* Used size_t for indexing throughout. Removed unused class

* Log improvements

* Resolved signed/unsigned warnings

* fix to path casing

* signed/unsigned adjustment

+++ PPPLib/include/PPPLib/Configuration/DarkCorrectionConfiguration.h
@@ -38,10 +38,10 @@ public:
         @param validFrom - the time from which this fit-window is valid
         @param validTo - the time to which this fit-window is valid
         @return 0 if sucessful, otherwise 1 */
-    int GetDarkSettings(int index, CDarkSettings& dSettings, novac::CDateTime& validFrom, novac::CDateTime& validTo) const;
+    int GetDarkSettings(size_t index, CDarkSettings& dSettings, novac::CDateTime& validFrom, novac::CDateTime& validTo) const;
 
     /** Gets the number of dark-current settings configured for this spectrometer */
-    int GetSettingsNum() const;
+    size_t GetSettingsNum() const;
 
 
 private:
